+++ .idea/compiler.xml
@@ -8,7 +8,6 @@
         <outputRelativeToContentRoot value="true" />
         <module name="maven-archiver" />
         <module name="Budget" />
-        <module name="stream-task" />
       </profile>
     </annotationProcessing>
     <bytecodeTargetLevel>
