--- .idea/libraries/Maven__org_codehaus_plexus_plexus_utils_1_5_8.xml
@@ -1,13 +1,0 @@
-<component name="libraryTable">
-  <library name="Maven: org.codehaus.plexus:plexus-utils:1.5.8">
-    <CLASSES>
-      <root url="jar://$MAVEN_REPOSITORY$/org/codehaus/plexus/plexus-utils/1.5.8/plexus-utils-1.5.8.jar!/" />
-    </CLASSES>
-    <JAVADOC>
-      <root url="jar://$MAVEN_REPOSITORY$/org/codehaus/plexus/plexus-utils/1.5.8/plexus-utils-1.5.8-javadoc.jar!/" />
-    </JAVADOC>
-    <SOURCES>
-      <root url="jar://$MAVEN_REPOSITORY$/org/codehaus/plexus/plexus-utils/1.5.8/plexus-utils-1.5.8-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
