--- .idea/libraries/Maven__org_projectlombok_lombok_1_18_22.xml
@@ -1,13 +1,0 @@
-<component name="libraryTable">
-  <library name="Maven: org.projectlombok:lombok:1.18.22">
-    <CLASSES>
-      <root url="jar://$MAVEN_REPOSITORY$/org/projectlombok/lombok/1.18.22/lombok-1.18.22.jar!/" />
-    </CLASSES>
-    <JAVADOC>
-      <root url="jar://$MAVEN_REPOSITORY$/org/projectlombok/lombok/1.18.22/lombok-1.18.22-javadoc.jar!/" />
-    </JAVADOC>
-    <SOURCES>
-      <root url="jar://$MAVEN_REPOSITORY$/org/projectlombok/lombok/1.18.22/lombok-1.18.22-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
