--- .idea/libraries/Maven__org_projectlombok_lombok_maven_plugin_1_18_20_0.xml
@@ -1,13 +1,0 @@
-<component name="libraryTable">
-  <library name="Maven: org.projectlombok:lombok-maven-plugin:1.18.20.0">
-    <CLASSES>
-      <root url="jar://$MAVEN_REPOSITORY$/org/projectlombok/lombok-maven-plugin/1.18.20.0/lombok-maven-plugin-1.18.20.0.jar!/" />
-    </CLASSES>
-    <JAVADOC>
-      <root url="jar://$MAVEN_REPOSITORY$/org/projectlombok/lombok-maven-plugin/1.18.20.0/lombok-maven-plugin-1.18.20.0-javadoc.jar!/" />
-    </JAVADOC>
-    <SOURCES>
-      <root url="jar://$MAVEN_REPOSITORY$/org/projectlombok/lombok-maven-plugin/1.18.20.0/lombok-maven-plugin-1.18.20.0-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
