--- .idea/libraries/Maven__org_sonatype_plexus_plexus_build_api_0_0_7.xml
@@ -1,13 +1,0 @@
-<component name="libraryTable">
-  <library name="Maven: org.sonatype.plexus:plexus-build-api:0.0.7">
-    <CLASSES>
-      <root url="jar://$MAVEN_REPOSITORY$/org/sonatype/plexus/plexus-build-api/0.0.7/plexus-build-api-0.0.7.jar!/" />
-    </CLASSES>
-    <JAVADOC>
-      <root url="jar://$MAVEN_REPOSITORY$/org/sonatype/plexus/plexus-build-api/0.0.7/plexus-build-api-0.0.7-javadoc.jar!/" />
-    </JAVADOC>
-    <SOURCES>
-      <root url="jar://$MAVEN_REPOSITORY$/org/sonatype/plexus/plexus-build-api/0.0.7/plexus-build-api-0.0.7-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
