+++ Budget/src/main/java/lt/codeacademy/BudgetProgram.java
@@ -1,7 +1,9 @@
 package lt.codeacademy;
 
+import lt.codeacademy.service.BudgetRun;
+
 public class BudgetProgram {
   public static void main(String[] args) {
-    BudgetRun.run();
+    Menu.run();
   }
 }
