+++ Budget/src/main/java/lt/codeacademy/service/BudgetService.java
@@ -6,19 +6,15 @@ import lt.codeacademy.model.RecordType;
 import java.util.ArrayList;
 
 public class BudgetService {
-  private Double balance = 0D;
-  private final ArrayList<Record> records = new ArrayList<>();
+  private static Double balance = 0D;
+  private static final ArrayList<Record> records = new ArrayList<>();
 
-  public void addRecord(Record record){
+  public static void addRecord(Record record) {
     records.add(record);
-    if (record.getRecordType().equals(RecordType.INCOME)){
-      balance += Double.parseDouble(record.getAmount());
-    } else {
-      balance -= Double.parseDouble(record.getAmount());
-    }
+    calculateBalance(record);
   }
 
-  public void printData(int menuChoice) {
+  public void printRecords(int menuChoice) {
     if (menuChoice == 3) {
       System.out.println("Pajamų išrašas");
       for (Record line : records) {
@@ -26,7 +22,7 @@ public class BudgetService {
           System.out.println(line);
         }
       }
-    } else if (menuChoice == 4){
+    } else if (menuChoice == 4) {
       System.out.println("Išlaidų išrašas");
       for (Record line : records) {
         if (line.getRecordType().equals(RecordType.EXPENSES)) {
@@ -36,12 +32,20 @@ public class BudgetService {
     }
   }
 
-  public void deleteRecord(Integer id){
+  public void deleteRecord(Integer id) {
     for (Record record : records) {
-      if(record.getId().equals(id)){
+      if (record.getId().equals(id)) {
         records.remove(record);
         break;
       }
+    }
+  }
+
+  private static void calculateBalance(Record record) {
+    if (record.getRecordType().equals(RecordType.INCOME)) {
+      balance += Double.parseDouble(record.getAmount());
+    } else {
+      balance -= Double.parseDouble(record.getAmount());
     }
   }
